Start refactoring & cleaning up
