+++ Model/Service/Product/IndexerStatusService.php
@@ -16,7 +16,7 @@
  * 3. Neither the name of the copyright holder nor the names of its contributors
  * may be used to endorse or promote products derived from this software without
  * specific prior written permission.
- *
+ * o
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -34,39 +34,79 @@
  *
  */
 
-namespace Nosto\Tagging\Model\Service;
+namespace Nosto\Tagging\Model\Service\Product;
 
-use Magento\Catalog\Api\Data\ProductInterface;
-use Magento\Store\Api\Data\StoreInterface;
-use Nosto\Object\Product\Product;
-use Nosto\Tagging\Model\Service\ProductServiceInterface as NostoProductService;
+use Magento\Framework\Indexer\IndexerRegistry;
+use Nosto\Tagging\Model\Mview\ChangeLogInterface;
+use Nosto\Tagging\Model\Mview\MviewInterface;
 
-class SanitizingProductService implements ProductServiceInterface
+class IndexerStatusService implements IndexerStatusServiceInterface
 {
-    /** @var ProductServiceInterface */
-    private $nostoProductService;
+    /** @var IndexerRegistry  */
+    private $indexerRegistry;
+
+    /** @var ChangeLogInterface */
+    private $changeLog;
+
+    /** @var MviewInterface */
+    private $mview;
 
     /**
-     * DefaultProductService constructor.
-     * @param ProductServiceInterface $nostoProductService
+     * @param ChangelogInterface $changeLog
+     * @param MviewInterface $mview
+     * @param IndexerRegistry $indexerRegistry
      */
     public function __construct(
-        NostoProductService $nostoProductService
+        ChangeLogInterface $changeLog,
+        MviewInterface $mview,
+        IndexerRegistry $indexerRegistry
     ) {
-        $this->nostoProductService = $nostoProductService;
+        $this->changeLog = $changeLog;
+        $this->mview = $mview;
+        $this->indexerRegistry = $indexerRegistry;
     }
 
     /**
      * @inheritDoc
      */
-    public function getProduct(ProductInterface $product, StoreInterface $store)
+    public function clearProcessedChangelog($indexerId)
     {
-        /** @var Product $nostoProduct */
-        $nostoProduct = $this->nostoProductService->getProduct(
-            $product,
-            $store
-        );
-        /** @noinspection PhpUnhandledExceptionInspection */
-        return $nostoProduct->sanitize();
+        if (!$this->isScheduled($indexerId)) {
+            return;
+        }
+        $this->mview->setId($indexerId);
+        $this->mview->clearChangelog();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getTotalChangelogCount($indexerId)
+    {
+        if (!$this->isScheduled($indexerId)) {
+            return 0;
+        }
+        $this->changeLog->setViewId($indexerId);
+        return $this->changeLog->getTotalRows();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getCurrentWatermark($indexerId)
+    {
+        if (!$this->isScheduled($indexerId)) {
+            return 0;
+        }
+        return (int)$this->mview->getState()->getVersionId();
+    }
+
+    /**
+     * @param $indexerId
+     * @return bool
+     */
+    private function isScheduled($indexerId)
+    {
+        return $this->indexerRegistry->get($indexerId)->isScheduled();
     }
 }
