+++ Model/Service/Product/SanitizingProductService.php
@@ -16,7 +16,7 @@
  * 3. Neither the name of the copyright holder nor the names of its contributors
  * may be used to endorse or promote products derived from this software without
  * specific prior written permission.
- * o
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -34,79 +34,39 @@
  *
  */
 
-namespace Nosto\Tagging\Model\Service;
+namespace Nosto\Tagging\Model\Service\Product;
 
-use Magento\Framework\Indexer\IndexerRegistry;
-use Nosto\Tagging\Model\Mview\ChangeLogInterface;
-use Nosto\Tagging\Model\Mview\MviewInterface;
+use Magento\Catalog\Api\Data\ProductInterface;
+use Magento\Store\Api\Data\StoreInterface;
+use Nosto\Object\Product\Product;
+use Nosto\Tagging\Model\Service\Product\ProductServiceInterface as NostoProductService;
 
-class IndexerStatusService implements IndexerStatusServiceInterface
+class SanitizingProductService implements ProductServiceInterface
 {
-    /** @var IndexerRegistry  */
-    private $indexerRegistry;
-
-    /** @var ChangeLogInterface */
-    private $changeLog;
-
-    /** @var MviewInterface */
-    private $mview;
+    /** @var ProductServiceInterface */
+    private $nostoProductService;
 
     /**
-     * @param ChangelogInterface $changeLog
-     * @param MviewInterface $mview
-     * @param IndexerRegistry $indexerRegistry
+     * DefaultProductService constructor.
+     * @param ProductServiceInterface $nostoProductService
      */
     public function __construct(
-        ChangeLogInterface $changeLog,
-        MviewInterface $mview,
-        IndexerRegistry $indexerRegistry
+        NostoProductService $nostoProductService
     ) {
-        $this->changeLog = $changeLog;
-        $this->mview = $mview;
-        $this->indexerRegistry = $indexerRegistry;
+        $this->nostoProductService = $nostoProductService;
     }
 
     /**
      * @inheritDoc
      */
-    public function clearProcessedChangelog($indexerId)
+    public function getProduct(ProductInterface $product, StoreInterface $store)
     {
-        if (!$this->isScheduled($indexerId)) {
-            return;
-        }
-        $this->mview->setId($indexerId);
-        $this->mview->clearChangelog();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getTotalChangelogCount($indexerId)
-    {
-        if (!$this->isScheduled($indexerId)) {
-            return 0;
-        }
-        $this->changeLog->setViewId($indexerId);
-        return $this->changeLog->getTotalRows();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getCurrentWatermark($indexerId)
-    {
-        if (!$this->isScheduled($indexerId)) {
-            return 0;
-        }
-        return (int)$this->mview->getState()->getVersionId();
-    }
-
-    /**
-     * @param $indexerId
-     * @return bool
-     */
-    private function isScheduled($indexerId)
-    {
-        return $this->indexerRegistry->get($indexerId)->isScheduled();
+        /** @var Product $nostoProduct */
+        $nostoProduct = $this->nostoProductService->getProduct(
+            $product,
+            $store
+        );
+        /** @noinspection PhpUnhandledExceptionInspection */
+        return $nostoProduct->sanitize();
     }
 }
